+++ libxode/src/sha.c
@@ -63,7 +63,7 @@ void shaUpdate(LIBXODE_SHA_CTX *ctx, unsigned char *dataIn, int len) {
    */
   for (i = 0; i < len; i++) {
     ctx->W[ctx->lenW / 4] <<= 8;
-    ctx->W[ctx->lenW / 4] |= (unsigned long)dataIn[i];
+    ctx->W[ctx->lenW / 4] |= (uint32_t)dataIn[i];
     if ((++ctx->lenW) % 64 == 0) {
       shaHashBlock(ctx);
       ctx->lenW = 0;
@@ -122,7 +122,7 @@ void shaBlock(unsigned char *dataIn, int len, unsigned char hashout[20]) {
 
 static void shaHashBlock(LIBXODE_SHA_CTX *ctx) {
   int t;
-  unsigned long A,B,C,D,E,TEMP;
+  uint32_t A,B,C,D,E,TEMP;
 
   for (t = 16; t <= 79; t++)
     ctx->W[t] =
